+++ protected/views/shared/_footer.php
@@ -1,0 +1,61 @@
+<footer class="base-footer-bar">
+    <div class="container">
+        <div class="row">
+            <div class="col-xs-4">
+                <ul class="list-inline base-footer-logo-bar">
+                    <li><a href="http://gigasciencepress.com/"><img src="/images/new_interface_image/gigascience_white.png" alt="Go to GigaScience Journal web site"></a></li>
+                </ul>
+            </div>
+            <div class="col-xs-4">
+                <ul class="base-footer-email">
+                    <li><a href="https://raw.githubusercontent.com/gigascience/gigadb-website/develop/CHANGELOG.md"><?php echo Yii::t('app' , 'Version: ' . Yii::app()->params["app_version"]) ?></a></li>
+                </ul>
+            </div>
+            <div class="col-xs-4 text-right">
+                <p class="base-footer-email"><a href="/site/contact"><i class="fa fa-envelope"></i> database@gigasciencejournal.com</a></p>
+                <ul class="list-inline base-footer-social-bar">
+                    <li>
+                        <a class="fa fa-facebook" href="http://facebook.com/GigaScience" title="GigaScience on Facebook" aria-label="GigaScience on Facebook"></a>
+                    </li>
+                    <li>
+                        <a class="fa fa-twitter" href="http://twitter.com/GigaScience" title="GigaScience on Twitter" aria-label="GigaScience on Twitter"></a>
+                    </li>
+                    <li class="icon-list-item">
+                        <a
+                          href="http://x.com/GigaScience"
+                          title="GigaScience on X"
+                          class="icon-list-item__link"
+                          aria-label="GigaScience on X"
+                        >
+                            <img
+                              class="icon-list-item__image"
+                              src="/images/icons/x-logo.svg"
+                              alt=""
+                            >
+                        </a>
+                    </li>
+                    <li>
+                        <a class="fa fa-weibo" href="http://weibo.com/gigasciencejournal" title="Gigascience on Weibo" aria-label="GigaScience on Weibo"></a>
+                    </li>
+                    <li class="icon-list-item">
+                        <a
+                          href="https://mastodon.social/@GigaScience"
+                          title="GigaScience on Mastodon"
+                          class="icon-list-item__link"
+                          aria-label="GigaScience on Mastodon"
+                        >
+                            <img
+                              class="icon-list-item__image"
+                              src="/images/icons/mastodon-logo.svg"
+                              alt=""
+                            >
+                        </a>
+                    </li>
+                    <li>
+                        <a class="fa fa-rss" href="http://gigasciencejournal.com/blog/" title="Gigascience Blog" aria-label="GigaScience Blog"></a>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </div>
+</footer>